chore: did a bunch of code cleanup and improvements
