+++ src/helper.ts
@@ -6,6 +6,22 @@ export function tryParseJson(content: string) {
   }
 }
 
+// export function merge<T extends Record<string, any>>(source: Partial<T>, target: Partial<T>): T {
+//   const merged = { ...source };
+//
+//   for (const key in target) {
+//     if (target[key] == null) continue;
+//
+//     if (key in merged && typeof target[key] === "object" && !Array.isArray(target[key])) {
+//       merged[key] = merge(source[key] as any, target[key] as any);
+//     } else {
+//       merged[key] = target[key] ?? source[key];
+//     }
+//   }
+//
+//   return merged as T;
+// }
+
 export function merge(source, target) {
   const merged = { ...source };
