+++ src/index.ts
@@ -1,6 +1,7 @@
 export { RabbitMQModule } from "./rabbitmq.module";
 export { RabbitMQService } from "./rabbitmq-service";
-export { AMQPConnectionManager } from "./amqp-connection-manager"
+export { AMQPConnectionManager } from "./amqp-connection-manager";
+export { RABBIT_OPTIONS } from "./rabbitmq.constants";
 export { ConfirmChannel, ConsumeMessage } from "amqplib";
 export {
   IRabbitMQConsumer,
